Move PauseIcon to components/icons
