+++ airbyte-webapp/src/components/icons/PauseIcon.tsx
